Corrected the error
When a message is received we only check if
the message is having the \n.When only \n is there then
variable line is  given a complete message. Else it won't assign the
message to the line. So the line is null is null when there is no \n.
so when the line is not equal to null then only it enter into
loop and print the messages.

But now the logic is changed such that it is splitted into multiple messages
depending on the \n. If there is no \n then it is taken as complete single message.